MYR-40 Address Claude Review: split converters.go + cover Unknown enum fallback
Claude Review flagged converters.go at 310 lines, over the CLAUDE.md
"Max 300 lines per file" rule. The new convertChargeState from the
decoder-wiring commit pushed the file over the cap.

Split: move the 8 enum → string converters (convertShiftState,
convertChargeState, convertDetailedChargeState, convertCarType,
convertSentryMode, convertDefrostMode, convertClimateKeeperMode,
convertHvacPower) into a new converters_enums.go. convertValue
dispatch, convertLocation, convertBool, convertNumericOrString,
convertRouteLine, convertTemperature, celsiusToFahrenheit, and
parseStringValue stay in converters.go.

Post-split line counts: converters.go 170, converters_enums.go 153
(both comfortably under the 300 cap).

Also covered Claude Review's optional suggestion: add an Unknown-
fallback test case to TestDecoder_DecodePayload_ChargeState_Proto2.
Passing tpb.ChargingState(9999) (an intentionally out-of-range enum
value) exercises the default branch of chargingStateString so a
future firmware release that adds an unrecognized enum value will
still produce a usable string on the wire.

No behavior change; pure mechanical refactor + one additional test.
